:recycle: Refactor to rename ProgressIndicatorView
Renamed to ProgressIndicator since it's the actual Widget
