+++ sample/src/main/java/org/smartregister/sample/presenter/SamplePresenter.java
@@ -1,7 +1,7 @@
 package org.smartregister.sample.presenter;
 
 import org.smartregister.reporting.contract.ReportContract;
-import org.smartregister.reporting.domain.BaseReportIndicatorsModel;
+import org.smartregister.reporting.model.BaseReportIndicatorsModel;
 import org.smartregister.reporting.domain.IndicatorQuery;
 import org.smartregister.reporting.domain.IndicatorTally;
 import org.smartregister.reporting.domain.ReportIndicator;
