+++ sample/src/main/res/layout/fragment_resources.xml
@@ -17,14 +17,14 @@
         android:layout_height="wrap_content"
         android:orientation="horizontal">
 
-        <org.smartregister.reporting.view.ProgressIndicatorView
+        <org.smartregister.reporting.view.ProgressIndicator
             android:id="@+id/progressIndicatorView"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_weight="1"
             app:subTitle="Sprayed" />
 
-        <org.smartregister.reporting.view.ProgressIndicatorView
+        <org.smartregister.reporting.view.ProgressIndicator
             android:id="@+id/progressIndicatorView2"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
